Made cast warnings go away. We should remove those static_casts, by using less diffrent types.

+++ Engine/HardDriveRoutines.cpp
@@ -103,7 +103,7 @@ void ReadSettPre2_4(ifstream &in, SimOptions &Options)
 	string line;
 	unsigned int x = 0;
 
-	Options.SpeciesNum = GrabNumber(in);
+	Options.SpeciesNum = static_cast<unsigned int>(GrabNumber(in));
 	Options.SpeciesNum++;
 
 	for (x = 0; x < Options.SpeciesNum; x++)
@@ -118,20 +118,21 @@ void ReadSettPre2_4(ifstream &in, SimOptions &Options)
 		/////////////////////////////////////////
 
 		
-		Options.Specie[x].Mutables.mutarray[0] = GrabNumber(in);
-		Options.Specie[x].path =				 GrabString(in);
-		Options.Specie[x].qty =					 GrabNumber(in);
-		Options.Specie[x].Name =				 GrabString(in);
-		Options.Specie[x].Veg =					 GrabBool(in);
-		Options.Specie[x].Fixed =				 GrabBool(in);
+		Options.Specie[x].Mutables.mutarray[0] =
+										static_cast<long int>(GrabNumber(in));
+		Options.Specie[x].path = GrabString(in);
+		Options.Specie[x].qty = static_cast<long unsigned int>(GrabNumber(in));
+		Options.Specie[x].Name = GrabString(in);
+		Options.Specie[x].Veg = static_cast<unsigned int>(GrabBool(in));
+		Options.Specie[x].Fixed = static_cast<unsigned int>(GrabBool(in));
 				
-		Options.Specie[x].color =				 GrabNumber(in);
-		line =									 GrabString(in);  //obsolete
-		Options.Specie[x].nrg =					 GrabNumber(in);
+		Options.Specie[x].color = static_cast<unsigned int>(GrabNumber(in));
+		line = GrabString(in); //obsolete
+		Options.Specie[x].nrg = static_cast<unsigned int>(GrabNumber(in));
 
 		for (unsigned int k = 0; k <= 13; k++)
 		{
-			Options.Specie[x].Mutables.mutarray[k] = GrabNumber(in);
+			Options.Specie[x].Mutables.mutarray[k] = static_cast<long int>(GrabNumber(in));
 		}
 
 		for (unsigned int l = 0; l <= 12; l++)
@@ -142,21 +143,21 @@ void ReadSettPre2_4(ifstream &in, SimOptions &Options)
 
 	Options.SimName =							GrabString(in);
 	line =										GrabString(in);  //obsolete
-	Options.FieldSize =							GrabNumber(in);
+	Options.FieldSize =							static_cast<unsigned int>(GrabNumber(in));
 	Options.FieldDimensions.set(				GrabNumber(in), GrabNumber(in), 0.0f);
-	Options.MaxPopulation =						GrabNumber(in);
+	Options.MaxPopulation =						static_cast<unsigned int>(GrabNumber(in));
 	line =										GrabString(in);  //obsolete
 	Options.DisableTies =						GrabBool(in);
-	Options.PopLimMethod =						GrabNumber(in);
+	Options.PopLimMethod =						static_cast<unsigned int>(GrabNumber(in));
 	Options.Toroidal =							GrabBool(in);
 	Options.NrgCyc =							GrabNumber(in);
-	Options.MinVegs =							GrabNumber(in);
+	Options.MinVegs =							static_cast<unsigned int>(GrabNumber(in));
 	Options.Costs[CONDCOST] =					GrabNumber(in);
 	Options.Costs[COSTSTORE] =					GrabNumber(in);
 	Options.Costs[SHOTCOST] =					GrabNumber(in);
 	Options.Costs[TIECOST] = Options.Costs[SHOTCOST];
 	Options.EnergyProp =						GrabNumber(in);
-	Options.EnergyFix =							GrabNumber(in);
+	Options.EnergyFix =							static_cast<unsigned int>(GrabNumber(in));
 	Options.EnergyExType =						GrabBool(in);
 
 	Options.YGravity =							GrabNumber(in);
@@ -166,14 +167,14 @@ void ReadSettPre2_4(ifstream &in, SimOptions &Options)
 	line =										GrabString(in); //obsolete
 
 	Options.AutoSimPath =						GrabString(in);
-	Options.AutoSimTime =						GrabNumber(in);
+	Options.AutoSimTime =						static_cast<unsigned int>(GrabNumber(in));
 	Options.AutoRobPath =						GrabString(in);
-	Options.AutoRobTime =						GrabNumber(in);
+	Options.AutoRobTime =						static_cast<unsigned int>(GrabNumber(in));
 
 	Options.MutCurrMult =						GrabNumber(in);
 	Options.MutOscill =							GrabBool(in);
-	Options.MutCycMax =							GrabNumber(in);
-	Options.MutCycMin =							GrabNumber(in);
+	Options.MutCycMax =							static_cast<long unsigned int>(GrabNumber(in));
+	Options.MutCycMin =							static_cast<long unsigned int>(GrabNumber(in));
 
 	Options.DBName =							GrabString(in);
 	Options.DBEnable =							GrabBool(in);
@@ -186,9 +187,9 @@ void ReadSettPre2_4(ifstream &in, SimOptions &Options)
 	if (!in.eof()) Options.Decay =				GrabNumber(in);
 	if (!in.eof()) Options.Gradient  =			GrabNumber(in);
 	if (!in.eof()) Options.DayNight  =			GrabBool(in);
-	if (!in.eof()) Options.CycleLength  =		GrabNumber(in);
-	if (!in.eof()) Options.DecayType  =			GrabNumber(in);
-	if (!in.eof()) Options.DecayDelay  =		GrabNumber(in);
+	if (!in.eof()) Options.CycleLength  =		static_cast<unsigned int>(GrabNumber(in));
+	if (!in.eof()) Options.DecayType  =			static_cast<unsigned int>(GrabNumber(in));
+	if (!in.eof()) Options.DecayDelay  =		static_cast<unsigned int>(GrabNumber(in));
 	if (!in.eof()) line =						GrabString(in); //obsolete
 
 	if (!in.eof()) Options.F1 =					GrabBool(in);
@@ -203,11 +204,11 @@ void ReadSettPre2_4(ifstream &in, SimOptions &Options)
 	}
 	if (!in.eof()) Options.Dxsxconnected  =		GrabBool(in);
 	if (!in.eof()) Options.Updnconnected  =		GrabBool(in);
-	if (!in.eof()) Options.RepopAmount  =		GrabNumber(in);
-	if (!in.eof()) Options.RepopCooldown  =		GrabNumber(in);
+	if (!in.eof()) Options.RepopAmount  =		static_cast<unsigned int>(GrabNumber(in));
+	if (!in.eof()) Options.RepopCooldown  =		static_cast<unsigned int>(GrabNumber(in));
 	if (!in.eof()) Options.ZeroMomentum  =		GrabBool(in);
 	if (!in.eof()) Options.UserSeedToggle  =	GrabBool(in);
-	if (!in.eof()) Options.UserSeedNumber  =	GrabNumber(in);
+	if (!in.eof()) Options.UserSeedNumber  =	static_cast<long int>(GrabNumber(in));
 	
 	
 	for (x = 0; x < Options.SpeciesNum; x++)
@@ -215,15 +216,16 @@ void ReadSettPre2_4(ifstream &in, SimOptions &Options)
 		for (int k = 14; k <= 20; k++)
 		{
 			if (!in.eof())
-				Options.Specie[x].Mutables.mutarray[k] = GrabNumber(in);
+				Options.Specie[x].Mutables.mutarray[k] = static_cast<long int>(GrabNumber(in));
 		}
 
 		if (!in.eof())
 				Options.Specie[x].Mutables.Mutations = GrabBool(in);
 	}
 	
-	if (!in.eof()) Options.VegFeedingMethod =	GrabNumber(in);
-	if (!in.eof()) Options.VegFeedingToBody =	GrabNumber(in); else Options.VegFeedingToBody = 0.1f;
+	if (!in.eof()) Options.VegFeedingMethod =	static_cast<unsigned int>(GrabNumber(in));
+	if (!in.eof()) Options.VegFeedingToBody =	GrabNumber(in);
+	else Options.VegFeedingToBody = 0.1f;
 	
 	if (!in.eof()) Options.CoefficientStatic =	GrabNumber(in);
 	if (!in.eof()) Options.CoefficientKinetic =	GrabNumber(in);
@@ -238,8 +240,8 @@ void ReadSettPre2_4(ifstream &in, SimOptions &Options)
 
 	for (x = 0; x < Options.SpeciesNum ; x++)
 	{
-		if (!in.eof()) Options.Specie[x].Mutables.CopyErrorWhatToChange = GrabNumber(in);
-		if (!in.eof()) Options.Specie[x].Mutables.PointWhatToChange = GrabNumber(in);
+		if (!in.eof()) Options.Specie[x].Mutables.CopyErrorWhatToChange = static_cast<unsigned int>(GrabNumber(in));
+		if (!in.eof()) Options.Specie[x].Mutables.PointWhatToChange = static_cast<unsigned int>(GrabNumber(in));
 
 		for (int k = 0; k <= 20; k++)
 		{
